+++ src/main/java/io/lunamc/protocol/packet/play/PlayPacketAllocator.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (C) 2017 LunaMC (https://lunamc.io/)
+ *
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+package io.lunamc.protocol.packet.play;
+
+public interface PlayPacketAllocator {
+}